move byName method from CommandDict to CommandParser
